Support downloading inventory to tempdir
This is needed because when running from the same remote [orchestration] host
although we specificy a new cloud to deploy on, since the location being checked on
remote is {{ansible_user_dir}}/ocpinv.json and it already exists from a previous
deployment, it could be used for the current deploy.

This patch defaults to using a temp location if orechestration is being done from a remote
host, and sticks to {{ playbook_dir }}/ocpinv.json when the localhost is the orchestration
host, for user convenience (if they wanted to manually edit the inventory file and have
JetSki consume it on a subsequent run).

Signed-off-by: Sai Sindhur Malleni <smalleni@redhat.com>