Set worker_count deployment fact
Userful when worker_count is not defined in group_vars/all.yml
in which case it defaults to all nodes except first 3 which are used
as masters. The worker_count var is already set if undefined on the
orchestration host, this just sets the fact on the provisioner as well.

Signed-off-by: Sai Sindhur Malleni <smalleni@redhat.com>