Removing Ansible as a package dep.
You can't run ansible without is already being installed. There are some
packaging dep conflicts in installing ansible from RPMs on RHEL in the
shared labs. This requirement will always hit that problem. Since
ansible has to be already installed to get to this point there's no
reason for it to be required here. If it's been installed by pip instead
of RPM then the deploy can't continue.